some landingpage cleanup and comments
