+++ controllers/landing_controller.js
@@ -11,10 +11,7 @@ router.get("/", function(req, res) {
     res.render("index", hbsObject);
 });
 
-router.get("/signin", function(req, res) {
-  var hbsObject = {};
-    res.render("signin", hbsObject);
-});
+
 
 router.get("/homepage", function(req, res) {
