--- controllers/skills_controller.js
@@ -1,17 +1,0 @@
-const express = require("express");
-const app = express();
-const path = require("path");
-const body = require("body-parser");
-const db = require(path.join(__dirname,".." ,"models"));
-
-//get route to pass skills wherever skills need to be displayed
-
-
-// module.exports = function(app) {
-
-// 	app.get("/", function(req, res) {
-// 	  var hbsObject = {};
-// 	    res.render("index", hbsObject);
-// 	});
-
-// };
