+++ server.js
@@ -47,8 +47,9 @@ app.use("/", routes);
 
 
 // Syncing our sequelize models and then starting our Express app
-//{ force: true }
+// { force: true }
 db.sequelize.sync().then(function() {
+
   app.listen(PORT, function() {
     console.log("App listening on PORT " + PORT);
   });
