+++ views/index.handlebars
@@ -1,23 +1,28 @@
 {{!-- This is the landing page --}}
 <div class="container">
   <div class="row" id="greeting">
+
     <div class="col-md-6">
-      <img src="/img/apple.svg" alt="landing-image">
+      <img src="/img/logo.png" alt="landing-image" id="landing-image">
     </div>
+
     <div class="col-md-6">
+    <div class="row" id="title_row">
       <div class="title">
-        <h2>Welcome to</h2>
         <h1>Student Skill Swap</h1>
-        <img id="logo_landing" src="img/logo.png" alt="logo">
-        <h4>An online platform that allows students to trade skills!</h4>
-          <button id="login_button" type="button" name="button">Login with Google</button>
+        <h4>Your chores are someone else's passion</h4>
+    </div>
+    <div class="row" id="button_row">
+          <button id="login_button" type="button" name="button"><img src="/img/google.png"></button>
+    </div>
       </div>
+
     </div>
   </div>
 </div>
 
 <div class="row creators">
-  <h3>The Creators:</h3>
+  <h3>Brought to you by:</h3>
   <div class="col-md-4">
     <img src="/img/sarah.jpg" class="coder" alt="Sarah's picture">
     <h4>Sarah</h4>
@@ -27,7 +32,7 @@
     <h4>Ashish</h4>
   </div>
   <div class="col-md-4">
-    <img src="/img/dp-placeholder.gif" class="coder" alt="Tali's picture">
+    <img src="/img/tali.png" class="coder" alt="Tali's picture">
     <h4>Tali</h4>
   </div>
 </div>
