+++ views/signin.handlebars
@@ -2,6 +2,9 @@
 {{!-- enter their offers and requests --}}
 
 
+{{!TALI: FISRT ONE IS SINGEL SELECT 2ND ONE IS MULTISELECT}}
+
+
 <div class="container">
   <form>
   Skill I need:<br>
